requirement.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
pydantic-settings==2.9.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pymysql==1.1.1
SQLAlchemy==2.0.41
email-validator==2.2.0